Pull out the MSSP check specific code from the telnet client
Prepping for a web client connection, which should be able to use the
base Telnet.Client
